api/src/requirements.txt
Django==4.2.5
django-environ==0.11.2
djangorestframework==3.14.0
gunicorn==21.2.0
pyTelegramBotAPI==4.13.0
psycopg2-binary==2.9.7